requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
pydantic-settings
numpy
transformers
torch
accelerate
